🔨 refactor: create getFromSchema tool
