+++ src/classes/Inventory.ts
@@ -4,7 +4,7 @@
 import { Effect, Paints, StrangeParts } from '../types/common';
 import SteamID from 'steamid';
 import TradeOfferManager, { EconItem, ItemAttributes } from 'steam-tradeoffer-manager';
-import SchemaManager, { Schema } from 'tf2-schema-2';
+import SchemaManager from 'tf2-schema-2';
 import SKU from 'tf2-sku-2';
 import Options from './Options';
 
@@ -12,7 +12,7 @@ import Bot from './Bot';
 // import log from '../lib/logger';
 
 import { noiseMakers, craftAll, uncraftAll } from '../lib/data';
-import { check } from '../lib/tools/export';
+import { check, getFromSchema } from '../lib/tools/export';
 
 export default class Inventory {
     private readonly steamID: SteamID;
@@ -127,15 +127,15 @@ export default class Inventory {
             items.filter(item => item.tradable),
             this.schema,
             this.options,
-            this.getPaints(this.schema),
-            this.getStrangeParts(this.schema)
+            getFromSchema.getPaints(this.schema),
+            getFromSchema.getStrangeParts(this.schema)
         );
         this.nonTradable = Inventory.createDictionary(
             items.filter(item => !item.tradable),
             this.schema,
             this.options,
-            this.getPaints(this.schema),
-            this.getStrangeParts(this.schema)
+            getFromSchema.getPaints(this.schema),
+            getFromSchema.getStrangeParts(this.schema)
         );
     }
 
@@ -231,82 +231,6 @@ export default class Inventory {
         return toObject;
     }
 
-    getStrangeParts(schema: Schema): StrangeParts {
-        const toObject: {
-            [name: string]: string;
-        } = {};
-
-        // Filter out built-in parts and also filter repeated "Kills"
-        const parts = schema.raw.schema.kill_eater_score_types.filter(
-            part =>
-                ![
-                    'Ubers',
-                    'Kill Assists',
-                    'Sentry Kills',
-                    'Sodden Victims',
-                    'Spies Shocked',
-                    'Heads Taken',
-                    'Humiliations',
-                    'Gifts Given',
-                    'Deaths Feigned',
-                    'Buildings Sapped',
-                    'Tickle Fights Won',
-                    'Opponents Flattened',
-                    'Food Items Eaten',
-                    'Banners Deployed',
-                    'Seconds Cloaked',
-                    'Health Dispensed to Teammates',
-                    'Teammates Teleported',
-                    'KillEaterEvent_UniquePlayerKills',
-                    'Points Scored',
-                    'Double Donks',
-                    'Teammates Whipped',
-                    'Wrangled Sentry Kills',
-                    'Carnival Kills',
-                    'Carnival Underworld Kills',
-                    'Carnival Games Won',
-                    'Contracts Completed',
-                    'Contract Points',
-                    'Contract Bonus Points',
-                    'Times Performed',
-                    'Kills and Assists during Invasion Event',
-                    'Kills and Assists on 2Fort Invasion',
-                    'Kills and Assists on Probed',
-                    'Kills and Assists on Byre',
-                    'Kills and Assists on Watergate',
-                    'Souls Collected',
-                    'Merasmissions Completed',
-                    'Halloween Transmutes Performed',
-                    'Power Up Canteens Used',
-                    'Contract Points Earned',
-                    'Contract Points Contributed To Friends'
-                ].includes(part.type_name) && ![0, 97].includes(part.type)
-        );
-
-        for (let i = 0; i < parts.length; i++) {
-            toObject[parts[i].type_name] = `sp${parts[i].type}`;
-        }
-
-        return toObject;
-    }
-
-    getPaints(schema: Schema): Paints {
-        const paintCans = schema.raw.schema.items.filter(
-            item => item.name.includes('Paint Can') && item.name !== 'Paint Can'
-        );
-        const toObject: {
-            [name: string]: string;
-        } = {};
-
-        for (let i = 0; i < paintCans.length; i++) {
-            if (paintCans[i].attributes === undefined) continue;
-
-            toObject[paintCans[i].item_name] = `p${paintCans[i].attributes[0].value}`;
-        }
-
-        return toObject;
-    }
-
     private static createDictionary(
         items: EconItem[],
         schema: SchemaManager.Schema,
@@ -397,11 +321,6 @@ export function getSkuAmountCanTrade(
     const amountCanTrade = bot.inventoryManager.amountCanTrade(sku, buying);
     const amountCanTradeGeneric = bot.inventoryManager.amountCanTrade(sku, buying, true);
     const mostCanTrade = amountCanTrade > amountCanTradeGeneric ? amountCanTrade : amountCanTradeGeneric;
-    const getUnusualEffects = () => {
-        return bot.schema.raw.schema.attribute_controlled_attached_particles.map(v => {
-            return { name: v.name, id: v.id };
-        });
-    };
     return {
         amountCanTradeGeneric: amountCanTradeGeneric,
         amountCanTrade: amountCanTrade,
@@ -409,6 +328,9 @@ export function getSkuAmountCanTrade(
         name:
             amountCanTrade > amountCanTradeGeneric
                 ? bot.schema.getName(SKU.fromString(sku))
-                : genericNameAndMatch(bot.schema.getName(SKU.fromString(sku), false), getUnusualEffects()).name
+                : genericNameAndMatch(
+                      bot.schema.getName(SKU.fromString(sku), false),
+                      getFromSchema.getUnusualEffects(bot.schema)
+                  ).name
     };
 }
