+++ src/classes/MyHandler/MyHandler.ts
@@ -45,7 +45,7 @@ import { exponentialBackoff } from '../../lib/helpers';
 
 import { craftAll, uncraftAll, giftWords, noiseMakers } from '../../lib/data';
 import { sendAlert, sendStats } from '../../lib/DiscordWebhook/export';
-import { summarize, check, uptime } from '../../lib/tools/export';
+import { summarize, check, uptime, getFromSchema } from '../../lib/tools/export';
 
 import genPaths from '../../resources/paths';
 
@@ -789,12 +789,11 @@ export default class MyHandler extends Handler {
 
             // Inform admin via Steam Chat or Discord Webhook Something Wrong Alert.
             const highValueOurNames: string[] = [];
-            const inventory = this.bot.inventoryManager.getInventory();
             const itemsName = check.getHighValueItems(
                 getHighValue.our.items,
                 this.bot,
-                inventory.getPaints(this.bot.schema),
-                inventory.getStrangeParts(this.bot.schema)
+                getFromSchema.getPaints(this.bot.schema),
+                getFromSchema.getStrangeParts(this.bot.schema)
             );
 
             if (opt.sendAlert.enable && opt.sendAlert.highValue.tryingToTake) {
